Fixed bug in double conversion: invalid parsing of 1.0000001 and similar values due to integer overflow in divisor. Float conversions seems to work fine with just int as the divisor.

+++ core/src/main/java/com/dslplatform/client/json/NumberConverter.java
@@ -236,7 +236,7 @@ public class NumberConverter {
 		if (i == len) return value;
 		else if (ch == '.') {
 			i++;
-			int div = 1;
+			long div = 1;
 			for (; i < len; i++) {
 				final int ind = buf[i] - 48;
 				div = (div << 3) + (div << 1);
@@ -265,7 +265,7 @@ public class NumberConverter {
 		if (i == len) return value;
 		else if (ch == '.') {
 			i++;
-			int div = 1;
+			long div = 1;
 			for (; i < len; i++) {
 				final int ind = buf[i] - 48;
 				div = (div << 3) + (div << 1);
